integration/requirements-test.txt
behave==1.2.6
docker==3.7.0
docker-compose==1.23.2
cryptography==2.3.0
argon2-cffi==16.3.0
Jinja2==2.10.1
pyhamcrest==1.9.0
PyNaCl==1.1.2
retrying==1.3.3
PGPy==0.4.3
flake8==3.6.0
flake8-commas==2.0.0
flake8-polyfill==1.0.2
flake8-print==3.1.0
flake8-quotes==1.0.0
flake8-snippets==0.2
flake8-string-format==0.2.3
pylint==2.1.1
yapf==0.24.0
bandit==1.5.1
mypy==0.660
mypy-extensions==0.4.1
